requirement.txt
sentence-transformers==3.4.1
langchain
flask
pypdf
python-dotenv
pinecone-client[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .
